requirements.txt
altgraph==0.15
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
entrypoints==0.2.3
Flask==1.0.2
Flask-JSGlue==0.3.1
Flask-Mail==0.9.1
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
keyring==13.2.1
macholib==1.9
MarkupSafe==1.0
modulegraph==0.16
oauthlib==2.1.0
Pillow==5.2.0
py2app==0.14
pyasn1==0.4.3
pyasn1-modules==0.2.2
python-dotenv==0.9.1
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
spotipy==2.4.4
tidalapi==0.5.0
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
whitenoise==3.3.1
yagmail==0.11.214

# requirements for Heroku
gunicorn==19.7.1